Patcher: AOB patterns: marked some original bytes as '?'
In some places, different original bytes are patched with the same patch
and grouped to the same result for efficiency. This means that the
source bytes are not always the same, which are now marked with '?'

Note that the source/original bytes are not actually used anywhere, this
is just for reference.